+++ src/plugins/contrib/symtab/symtabexec.cpp
@@ -505,7 +505,7 @@ int SymTabExecDlg::ParseOutputSuccess(wxString lib, wxString filter)
     return 0;
   }
 
-	Manager::Get()->GetLogManager()->DebugLog(F(_T("SymTab: Parsing %d items..."), count));
+	Manager::Get()->GetLogManager()->DebugLog(F(_T("SymTab: Parsing %lu items..."), static_cast<unsigned long>(count)));
 
   wxProgressDialog* progress = 0L;
   if (count>2000) // avoid flickering for small libs
