+++ src/plugins/contrib/wxSmithContribItems/wxSmithKWIC/wxsbmpswitcher.cpp
@@ -98,7 +98,7 @@ void wxsBmpSwitcher::OnBuildCreatingCode()
 			}
 			// Default is 0. If state <= 0 or > number of bitmaps, use default.
 			if(m_iState > 0 && m_iState < (long)m_arrBmps.Count()){
-				Codef(_T("\t%ASetState(%d);\n"), m_iState);
+				Codef(_T("\t%ASetState(%ld);\n"), m_iState);
 			}
 
 			BuildSetupWindowCode();
@@ -266,7 +266,7 @@ bool wxsBmpSwitcher::OnXmlWrite(TiXmlElement *Element, bool IsXRC, bool IsExtra)
 
     for(size_t i = 0;i < m_arrBmps.Count();i++){
         BmpDesc *Desc = m_arrBmps[i];
-        wxString s = wxString::Format(wxT("bitmap_%d"), i + 1);
+        wxString s = wxString::Format(wxT("bitmap_%lu"), static_cast<unsigned long>(i + 1));
         TiXmlElement *msg = new TiXmlElement(s.mb_str());
         msg->LinkEndChild(new TiXmlText(Desc->sPath.mb_str()));
         tags->LinkEndChild(msg);
